% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covid_eol}
\alias{covid_eol}
\title{Synthetic ICU end-of-life PTSD survey data}
\format{
A data frame with 298 rows and 16 variables:
\describe{
\item{subject_id}{Integer. Unique subject identifier.}
\item{ies_total}{Integer. IES-6 total score (0--24).}
\item{age}{Integer. Respondent age in years.}
\item{ethnicity}{Factor with levels \code{Hispanic}, \code{Non-Hispanic}.}
\item{race}{Factor with levels \code{AAPI/NA/NI}, \code{Black}, \code{Other}, \code{White}.}
\item{sex}{Factor with levels \code{Female}, \code{Male}.}
\item{education}{Factor with levels \code{<= High school diploma or GED},
\code{4-year college degree}, \code{Some college or trade school},
\code{Some graduate school or graduate degree}.}
\item{relationship}{Factor with levels \code{Child}, \code{Other},
\code{Parent}, \code{Sibling}, \code{Spouse/partner}.
Relationship of the respondent to the ICU patient.}
\item{resp_failure}{Factor with levels \code{High Flow nasal cannula},
\code{Invasive MV}, \code{nasal cannula}, \code{Non-invasive MV}.
Type of respiratory failure treatment.}
\item{ecmo}{Factor with levels \code{No}, \code{Yes}.
Whether extracorporeal membrane oxygenation was used.}
\item{icu_totaldays}{Integer. Total days in the ICU.}
\item{death}{Logical. Whether the ICU patient died.}
\item{psych_meds}{Factor with levels \code{No}, \code{Yes}.
Whether the respondent was taking psychiatric medications.}
\item{charlson}{Integer. Charlson Comorbidity Index score.}
\item{sofa}{Integer. Sequential Organ Failure Assessment score (0--20).}
\item{ICU_admit_month}{Factor with levels \code{April},
\code{Feburary/March}, \code{May/June/July}.
Month of ICU admission.}
}
}
\source{
Synthetic data generated to resemble a real ICU family-member survey.
Not from actual patient records.
}
\usage{
covid_eol
}
\description{
Synthetic data modeled after a 2021--2022 study of family members of
ICU patients who were isolated from loved ones during the COVID-19 pandemic.
Respondents were assessed for PTSD using the Impact of Events Scale-6
(IES-6); a score of 10 or higher indicates likely PTSD.
}
\keyword{datasets}
